test(relay): streamline plugin test modules
